requirements.txt
numpy==1.13.0
PyYAML==3.12
tensorflow-gpu==1.4.0
